Break down `types`
