+++ web/client-js/src/store/board/actions/create-board-action.ts
@@ -24,21 +24,18 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { newBoardId } from "@/store/board/id-factory";
 import { BoardAction } from "@/store/board/actions/base/board-action";
+import { newBoardId } from "@/store/board/id-factory";
 import { client } from "@/dependency/container";
-import { ActionContext, ActionHandler } from "vuex";
-import {
-  BoardCreated,
-  BoardState,
-  ColumnAdded,
-  Mutation,
-} from "@/store/board/types";
-import { RootState } from "@/store/root/types";
+import { MutationType } from "@/store/board/mutations";
 import { Event } from "spine-web/proto/spine/core/event_pb";
+import { BoardCreated, ColumnAdded } from "@/store/board/common";
 import { AnyPacker } from "spine-web/client/any-packer";
 import { Type } from "spine-web/client/typed-message";
 import { Filters } from "spine-web";
+import { ActionContext, ActionHandler } from "vuex";
+import { RootState } from "@/store/root/root-state";
+import { BoardState } from "@/store/board/state/board-state";
 
 type CreateBoard = proto.spine_examples.kanban.CreateBoard;
 type BoardId = proto.spine_examples.kanban.BoardId;
@@ -78,7 +75,10 @@ export default class CreateBoardAction extends BoardAction<null, void> {
           const innerEvent: BoardCreated = AnyPacker.unpack(e.getMessage()).as(
             Type.forClass(proto.spine_examples.kanban.BoardCreated)
           );
-          this.getActionContext().commit(Mutation.BOARD_CREATED, innerEvent);
+          this.getActionContext().commit(
+            MutationType.BOARD_CREATED,
+            innerEvent
+          );
         });
       });
   }
@@ -97,7 +97,7 @@ export default class CreateBoardAction extends BoardAction<null, void> {
           const innerEvent: BoardCreated = AnyPacker.unpack(e.getMessage()).as(
             Type.forClass(proto.spine_examples.kanban.ColumnAdded)
           );
-          this.getActionContext().commit(Mutation.COLUMN_ADDED, innerEvent);
+          this.getActionContext().commit(MutationType.COLUMN_ADDED, innerEvent);
         });
       });
   }
