+++ web/client-js/src/store/board/actions/index.ts
@@ -25,17 +25,31 @@
  */
 
 import { ActionTree } from "vuex";
-import { Action, BoardState } from "@/store/board/types";
-import { RootState } from "@/store/root/types";
+import { BoardState } from "@/store/board/state/board-state";
+import { RootState } from "@/store/root/root-state";
 import CreateBoardAction from "@/store/board/actions/create-board-action";
 import AddColumnAction from "@/store/board/actions/add-column-action";
 
 /**
- * Exposes actions to interact with the remote board state.
+ * Defines action types to interact with the remote board state.
  */
-const actions: ActionTree<BoardState, RootState> = {
-  [Action.CREATE_BOARD]: CreateBoardAction.newHandler(),
-  [Action.ADD_COLUMN]: AddColumnAction.newHandler(),
+export const ActionType = {
+  /**
+   * Subscribes to the {@link BoardCreated} and {@link ColumnAdded} events and sends
+   * the {@code proto.spine_examples.kanban.CreateBoard} command to create a board.
+   */
+  CREATE_BOARD: "createBoard",
+
+  /**
+   * Sends the {@code proto.spine_examples.kanban.AddColumn} command to add the column.
+   */
+  ADD_COLUMN: "addColumn",
 };
 
-export default actions;
+/**
+ * Exposes actions to interact with the remote board state.
+ */
+export const actions: ActionTree<BoardState, RootState> = {
+  [ActionType.CREATE_BOARD]: CreateBoardAction.newHandler(),
+  [ActionType.ADD_COLUMN]: AddColumnAction.newHandler(),
+};
