+++ web/client-js/src/store/board/mutations/board-created-mutation.ts
@@ -25,7 +25,8 @@
  */
 
 import { Mutation } from "vuex";
-import { BoardCreated, BoardState } from "@/store/board/types";
+import { BoardState } from "@/store/board/state/board-state";
+import { BoardCreated } from "@/store/board/common";
 
 /**
  * Mutates the local {@linkplain BoardState board state} in response
